Move likePatternToRegexp from Functions/ to Common/
