+++ src/Access/Common/AllowedClientHosts.cpp
@@ -1,8 +1,8 @@
 #include <Access/Common/AllowedClientHosts.h>
 #include <Common/Exception.h>
+#include <Common/likePatternToRegexp.h>
 #include <Common/logger_useful.h>
 #include <base/scope_guard.h>
-#include <Functions/likePatternToRegexp.h>
 #include <Poco/Net/SocketAddress.h>
 #include <Poco/RegularExpression.h>
 #include <boost/algorithm/string/predicate.hpp>
