+++ src/Analyzer/Passes/ConvertOrLikeChainPass.cpp
@@ -4,12 +4,13 @@
 #include <unordered_map>
 #include <vector>
 
+#include <Common/likePatternToRegexp.h>
+
 #include <Core/Field.h>
 
 #include <DataTypes/DataTypesNumber.h>
 
 #include <Functions/FunctionFactory.h>
-#include <Functions/likePatternToRegexp.h>
 
 #include <Interpreters/Context.h>
