+++ src/Common/likePatternToRegexp.h
