+++ src/Interpreters/ConvertFunctionOrLikeVisitor.cpp
@@ -1,9 +1,9 @@
-#include <Functions/likePatternToRegexp.h>
 #include <Interpreters/ConvertFunctionOrLikeVisitor.h>
 #include <Parsers/ASTFunction.h>
 #include <Parsers/ASTIdentifier.h>
 #include <Parsers/ASTLiteral.h>
 #include <Parsers/IAST.h>
+#include <Common/likePatternToRegexp.h>
 #include <Common/typeid_cast.h>
